Cleanup a bit of dead code
