+++ lib/client.js
@@ -6,9 +6,9 @@
  * README.md file in the root directory of this source tree.
  */
 
-var crypto = require('crypto');
 var EventEmitter = require('events').EventEmitter;
 var util = require('util');
+var utils = require('./utils')
 var pgPass = require('pgpass');
 var TypeOverrides = require('./type-overrides');
 
@@ -97,8 +97,8 @@ Client.prototype.connect = function(callback) {
 
   //password request handling
   con.on('authenticationMD5Password', checkPgPass(function(msg) {
-    var inner = Client.md5(self.password + self.user);
-    var outer = Client.md5(Buffer.concat([Buffer.from(inner), msg.salt]));
+    var inner = utils.md5(self.password + self.user);
+    var outer = utils.md5(Buffer.concat([Buffer.from(inner), msg.salt]));
     var md5password = "md5" + outer;
     con.password(md5password);
   }));
@@ -404,10 +404,6 @@ Client.prototype.end = function (cb) {
       this.connection.once('end', resolve)
     })
   }
-};
-
-Client.md5 = function (string) {
-  return crypto.createHash('md5').update(string, 'utf-8').digest('hex');
 };
 
 // expose a Query constructor
