+++ lib/index.js
@@ -12,7 +12,22 @@ var Client = require('./client');
 var defaults =  require('./defaults');
 var Connection = require('./connection');
 var ConnectionParameters = require('./connection-parameters');
-var poolFactory = require('./pool-factory');
+var Pool = require('pg-pool');
+
+const poolFactory = (Client) => {
+  var BoundPool = function(options) {
+    var config = { Client: Client };
+    for (var key in options) {
+      config[key] = options[key];
+    }
+    Pool.call(this, config);
+  };
+
+  util.inherits(BoundPool, Pool);
+
+  return BoundPool;
+};
+
 
 var PG = function(clientConstructor) {
   this.defaults = defaults;
