+++ lib/native/index.js
@@ -1,1 +1,1 @@
-module.exports = require('./client');
+module.exports = require('./client')
