+++ lib/utils.js
@@ -7,7 +7,9 @@
  */
 
 const util = require('util')
-var defaults = require('./defaults');
+const crypto = require('crypto');
+
+const defaults = require('./defaults');
 
 function escapeElement(elementRepresentation) {
   var escaped = elementRepresentation
@@ -138,6 +140,11 @@ function normalizeQueryConfig (config, values, callback) {
   return config;
 }
 
+
+const md5 = function (string) {
+  return crypto.createHash('md5').update(string, 'utf-8').digest('hex');
+};
+
 module.exports = {
   prepareValue: function prepareValueWrapper (value) {
     //this ensures that extra arguments do not get passed into prepareValue
@@ -145,4 +152,5 @@ module.exports = {
     return prepareValue(value);
   },
   normalizeQueryConfig: normalizeQueryConfig,
+  md5: md5,
 };
