+++ test/unit/client/md5-password-tests.js
@@ -1,4 +1,5 @@
-require(__dirname + '/test-helper');
+require('./test-helper');
+var utils = require('../../../lib/utils')
 
 test('md5 authentication', function() {
   var client = createClient();
@@ -9,8 +10,8 @@ test('md5 authentication', function() {
   test('responds', function() {
     assert.lengthIs(client.connection.stream.packets, 1);
     test('should have correct encrypted data', function() {
-      var encrypted = Client.md5(client.password + client.user);
-      encrypted = Client.md5(encrypted + salt.toString('binary'));
+      var encrypted = utils.md5(client.password + client.user);
+      encrypted = utils.md5(encrypted + salt.toString('binary'));
       var password = "md5" + encrypted
       //how do we want to test this?
       assert.equalBuffers(client.connection.stream.packets[0], new BufferList()
@@ -20,5 +21,5 @@ test('md5 authentication', function() {
 });
 
 test('md5 of utf-8 strings', function() {
-  assert.equal(Client.md5('😊'), '5deda34cd95f304948d2bc1b4a62c11e');
+  assert.equal(utils.md5('😊'), '5deda34cd95f304948d2bc1b4a62c11e');
 });
